backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.1
fastapi==0.127.0
fastjsonschema==2.21.2
filelock==3.20.1
fsspec==2025.12.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
mpmath==1.3.0
nbclient==0.10.3
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
webencodings==0.5.1
yarg==0.1.9
